pyproject.toml
[tool.poetry]
name = "strawberry-django-playground"
version = "0.1.0"
description = ""
authors = ["ronald.williams <ronald.williams@onepeloton.com>"]
readme = "README.md"
packages = [{include = "strawberry_django_playground"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
strawberry-graphql = "^0.133.5"
strawberry-graphql-django = "^0.6"
strawberry-django-plus = "^1.28.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

